remove unless function and merge files
