--- test_utils.go
@@ -1,53 +1,0 @@
-// Copyright 2013 Hui Chen
-// Copyright 2016 ego authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License"): you may
-// not use this file except in compliance with the License. You may obtain
-// a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-// License for the specific language governing permissions and limitations
-// under the License.
-
-package gse
-
-import (
-	"fmt"
-	"testing"
-)
-
-func printTokens(tokens []*Token, numTokens int) (output string) {
-	for iToken := 0; iToken < numTokens; iToken++ {
-		for _, word := range tokens[iToken].text {
-			output += fmt.Sprint(string(word))
-		}
-		output += " "
-	}
-	return
-}
-
-func toWords(strings ...string) []Text {
-	words := []Text{}
-	for _, s := range strings {
-		words = append(words, []byte(s))
-	}
-	return words
-}
-
-func bytesToString(bytes []Text) (output string) {
-	for _, b := range bytes {
-		output += (string(b) + "/")
-	}
-	return
-}
-
-func expect(t *testing.T, expect string, actual interface{}) {
-	actualString := fmt.Sprint(actual)
-	if expect != actualString {
-		t.Errorf("期待值=\"%s\", 实际=\"%s\"", expect, actualString)
-	}
-}
